hack around with project structure, starting to look better and better
