+++ auth.py
@@ -1,0 +1,19 @@
+from flask import Blueprint
+from . import mongo
+
+auth = Blueprint("auth", __name__)
+
+
+@auth.route('/login')
+def login():
+    return 'Login'
+
+
+@auth.route('/signup')
+def signup():
+    return 'Signup'
+
+
+@auth.route('/logout')
+def logout():
+    return 'Logout'
