+++ install_dependencies.sh
@@ -10,3 +10,5 @@ python -m pip install dnspython
 python -m pip install graphene
 python -m pip install Flask
 python -m pip install Flask_Pymongo
+python -m pip install Flask-Login
+python -m pip install bcrypt
