+++ run.sh
@@ -1,0 +1,3 @@
+export FLASK_APP=.
+export FLASK_DEBUG=1
+flask run
